DRFL-EAAI2023/requirements.txt
absl-py==0.11.0
albumentations==1.1.0
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
cachetools==4.2.1
certifi==2020.12.5
cffi==1.13.1
chardet==3.0.4
contextlib2==0.6.0.post1
cupy-cuda101==9.0.0
cycler==0.10.0
dataclasses==0.8
decorator==4.4.1
defusedxml==0.6.0
docopt==0.6.2
dominate==2.4.0
dropblock==0.3.0
easydict==1.9
efficientnet-pytorch==0.6.3
einops==0.3.0
entrypoints==0.3
fastrlock==0.6
future==0.18.2
fvcore==0.1.5.post20210630
google-auth==1.27.0
google-auth-oauthlib==0.4.2
grpcio==1.35.0
h5py==2.10.0
idna==2.8
imageio==2.6.1
importlib-metadata==1.7.0
install==1.3.4
iopath==0.1.9
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.14.1
jsonpatch==1.25
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-core==4.6.3
kiwisolver==1.1.0
llvmlite==0.31.0
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.1.1
MedPy==0.4.0
mistune==0.8.4
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
ml-collections==0.1.0
munch==2.5.0
natsort==7.0.1
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
nibabel==3.0.1
notebook==6.0.3
numba==0.48.0
numpy==1.17.2
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.5.62
opencv-python-headless==4.5.5.62
packaging==20.4
pandas==1.0.0
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.2.0
portalocker==2.3.0
pretrainedmodels==0.7.4
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.10.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycococreatortools==0.2.0
pycparser==2.19
pydensecrf==1.0rc3
Pygments==2.6.1
pykwalify==1.7.0
pyparsing==2.4.2
pyradiomics==2.2.0
pyrsistent==0.16.0
python-dateutil==2.8.0
pytz==2019.3
PyWavelets==1.0.0
PyYAML==5.1.2
pyzmq==18.1.1
qudida==0.0.4
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.7.1
scikit-image==0.16.2
scikit-learn==0.19.1
scipy==1.1.0
seaborn==0.11.2
segmentation-models-pytorch==0.2.1
Send2Trash==1.5.0
SimpleITK==1.2.3
six==1.12.0
sklearn==0.0
tabulate==0.8.9
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboardX==1.9
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
timm==0.4.12
torch==1.6.0
torchfile==0.1.0
torchsummary==1.5.1
torchvision==0.7.0
tornado==6.0.3
tqdm==4.19.9
traitlets==4.3.3
typing-extensions==4.0.1
urllib3==1.25.8
visdom==0.1.8.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
yacs==0.1.8
zipp==3.1.0
